Extract compose codeviewer into separate module
