+++ compose/ui/api/ui.api
@@ -2,8 +2,8 @@ public final class io/github/composegears/valkyrie/compose/ui/InfoCardKt {
 	public static final fun InfoCard (Lkotlin/jvm/functions/Function0;Landroidx/compose/ui/graphics/vector/ImageVector;Ljava/lang/String;Ljava/lang/String;Landroidx/compose/ui/Modifier;Landroidx/compose/runtime/Composer;II)V
 }
 
-public final class io/github/composegears/valkyrie/compose/ui/util/ColorKt {
-	public static final fun dim-8_81llA (J)J
-	public static final fun disabled-8_81llA (J)J
+public final class io/github/composegears/valkyrie/compose/ui/foundation/ScrollbarKt {
+	public static final fun HorizontalScrollbar (Landroidx/compose/foundation/layout/BoxScope;Landroidx/compose/foundation/v2/ScrollbarAdapter;Landroidx/compose/ui/Modifier;Landroidx/compose/runtime/Composer;II)V
+	public static final fun VerticalScrollbar (Landroidx/compose/foundation/layout/BoxScope;Landroidx/compose/foundation/v2/ScrollbarAdapter;Landroidx/compose/ui/Modifier;Landroidx/compose/runtime/Composer;II)V
 }
