+++ compose/ui/api/ui.klib.api
@@ -6,6 +6,6 @@
 // - Show declarations: true
 
 // Library unique name: <valkyrie.compose:ui>
-final fun (androidx.compose.ui.graphics/Color).io.github.composegears.valkyrie.compose.ui.util/dim(): androidx.compose.ui.graphics/Color // io.github.composegears.valkyrie.compose.ui.util/dim|dim@androidx.compose.ui.graphics.Color(){}[0]
-final fun (androidx.compose.ui.graphics/Color).io.github.composegears.valkyrie.compose.ui.util/disabled(): androidx.compose.ui.graphics/Color // io.github.composegears.valkyrie.compose.ui.util/disabled|disabled@androidx.compose.ui.graphics.Color(){}[0]
+final fun (androidx.compose.foundation.layout/BoxScope).io.github.composegears.valkyrie.compose.ui.foundation/HorizontalScrollbar(androidx.compose.foundation.v2/ScrollbarAdapter, androidx.compose.ui/Modifier?, androidx.compose.runtime/Composer?, kotlin/Int, kotlin/Int) // io.github.composegears.valkyrie.compose.ui.foundation/HorizontalScrollbar|HorizontalScrollbar@androidx.compose.foundation.layout.BoxScope(androidx.compose.foundation.v2.ScrollbarAdapter;androidx.compose.ui.Modifier?;androidx.compose.runtime.Composer?;kotlin.Int;kotlin.Int){}[0]
+final fun (androidx.compose.foundation.layout/BoxScope).io.github.composegears.valkyrie.compose.ui.foundation/VerticalScrollbar(androidx.compose.foundation.v2/ScrollbarAdapter, androidx.compose.ui/Modifier?, androidx.compose.runtime/Composer?, kotlin/Int, kotlin/Int) // io.github.composegears.valkyrie.compose.ui.foundation/VerticalScrollbar|VerticalScrollbar@androidx.compose.foundation.layout.BoxScope(androidx.compose.foundation.v2.ScrollbarAdapter;androidx.compose.ui.Modifier?;androidx.compose.runtime.Composer?;kotlin.Int;kotlin.Int){}[0]
 final fun io.github.composegears.valkyrie.compose.ui/InfoCard(kotlin/Function0<kotlin/Unit>, androidx.compose.ui.graphics.vector/ImageVector, kotlin/String, kotlin/String, androidx.compose.ui/Modifier?, androidx.compose.runtime/Composer?, kotlin/Int, kotlin/Int) // io.github.composegears.valkyrie.compose.ui/InfoCard|InfoCard(kotlin.Function0<kotlin.Unit>;androidx.compose.ui.graphics.vector.ImageVector;kotlin.String;kotlin.String;androidx.compose.ui.Modifier?;androidx.compose.runtime.Composer?;kotlin.Int;kotlin.Int){}[0]
