+++ compose/util/api/util.api
@@ -1,0 +1,9 @@
+public final class io/github/composegears/valkyrie/compose/util/ColorKt {
+	public static final fun dim-8_81llA (J)J
+	public static final fun disabled-8_81llA (J)J
+}
+
+public final class io/github/composegears/valkyrie/compose/util/ColorSchemeKt {
+	public static final fun isLight (Landroidx/compose/material3/ColorScheme;Landroidx/compose/runtime/Composer;I)Z
+}
+
