+++ compose/util/build.gradle.kts
@@ -6,9 +6,6 @@ plugins {
 kotlin {
     sourceSets {
         commonMain.dependencies {
-            implementation(projects.compose.core)
-            implementation(projects.compose.util)
-
             implementation(compose.material3)
         }
     }
