+++ compose/util/src/commonMain/kotlin/io/github/composegears/valkyrie/compose/util/ColorScheme.kt
@@ -1,0 +1,11 @@
+package io.github.composegears.valkyrie.compose.util
+
+import androidx.compose.material3.ColorScheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Stable
+import androidx.compose.ui.graphics.luminance
+
+@Stable
+val ColorScheme.isLight
+    @Composable
+    get() = background.luminance() > 0.5f
