+++ tools/compose-app/src/commonMain/kotlin/io/github/composegears/valkyrie/screen/intro/IntroScreen.kt
@@ -22,7 +22,7 @@ import io.github.composegears.valkyrie.compose.core.layout.WeightSpacer
 import io.github.composegears.valkyrie.compose.icons.ValkyrieIcons
 import io.github.composegears.valkyrie.compose.icons.outlined.Conversion
 import io.github.composegears.valkyrie.compose.ui.InfoCard
-import io.github.composegears.valkyrie.compose.ui.util.dim
+import io.github.composegears.valkyrie.compose.util.dim
 import io.github.composegears.valkyrie.screen.mode.simple.SimpleConversionScreen
 import io.github.composegears.valkyrie.shared.Mode
 import org.jetbrains.compose.resources.stringResource
