+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/foundation/theme/Theme.kt
@@ -3,18 +3,15 @@ package io.github.composegears.valkyrie.ui.foundation.theme
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.ColorScheme
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.Stable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.luminance
 import androidx.compose.ui.platform.LocalInspectionMode
 import com.intellij.openapi.project.Project
 import io.github.composegears.valkyrie.ui.foundation.compositionlocal.LocalProject
@@ -145,8 +142,3 @@ private fun PreviewWrapper(content: @Composable () -> Unit) {
         content = content,
     )
 }
-
-@Stable
-val ColorScheme.isLight
-    @Composable
-    get() = background.luminance() > 0.5f
